Fix: statistic fixed
