+++ BackEnd/admin/src/Routes.js
@@ -1,12 +1,12 @@
 import React from "react";
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
-import UserStaticsPage from "./UserStaticsPage/";
+import UserstatisticPage from "./UserstatisticPage/";
 
 const Routes = () => {
   return (
     <Router basename={process.env.REACT_APP_BASE}>
       <Switch>
-        <Route exact path="/" component={UserStaticsPage} />
+        <Route exact path="/" component={UserstatisticPage} />
       </Switch>
     </Router>
   );
