+++ BackEnd/app/src/routes/index.js
@@ -1,13 +1,13 @@
 import { Router } from 'express';
 
-import StaticsRouter from './statics.js';
+import statisticRouter from './statistic.js';
 import UserRouter from './user.js';
 import DuelRouter from './duel.js';
 
 export default () => {
   const router = Router();
 
-  router.use('/statics', StaticsRouter);
+  router.use('/statistic', statisticRouter);
   router.use('/users', UserRouter);
   router.use('/duels', DuelRouter);
